feat: add datetime to log
