+++ api/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 rocket = { version = "0.5.0-rc.1", features = ["uuid", "serde_json", "json"] }
 cargo = "0.59.0"
 cargo-util = "0.1"
+chrono = "0.4"
 anyhow = "1.0.54"
 libloading = "0.7.3"
 futures = "0.3"
