+++ cargo-shuttle/Cargo.toml
@@ -8,7 +8,10 @@ description = "A cargo command for the shuttle platform (https://www.shuttle.rs/
 
 [dependencies]
 anyhow = "1.0.53"
+chrono = "0.4"
+colored = "2.0"
 dirs = "4.0.0"
+log = "0.4"
 serde_json = "1.0.79"
 serde = { version = "1.0.136", features = ["derive"] }
 reqwest = { version = "0.11.9", features = ["json"] }
