+++ cargo-shuttle/src/client.rs
@@ -1,4 +1,7 @@
 use anyhow::{anyhow, Context, Result};
+use chrono::{DateTime, Local};
+use colored::{ColoredString, Colorize};
+use log::Level;
 use reqwest::{Response, StatusCode};
 use reqwest_middleware::{ClientBuilder, ClientWithMiddleware};
 use reqwest_retry::{policies::ExponentialBackoff, RetryTransientMiddleware};
@@ -68,11 +71,30 @@ pub(crate) async fn logs(api_url: ApiUrl, api_key: &ApiKey, project: &ProjectNam
 
     let deployment_meta = get_deployment_meta(api_url, api_key, project, &client).await?;
 
-    for log in deployment_meta.runtime_logs {
-        println!("{}", log);
+    for (datetime, log) in deployment_meta.runtime_logs {
+        let datetime: DateTime<Local> = DateTime::from(datetime);
+        println!(
+            "{}{} {:<5} {}{} {}",
+            "[".bright_black(),
+            datetime.format("%Y-%m-%dT%H:%M:%SZ"),
+            get_colored_level(&log.level),
+            log.target,
+            "]".bright_black(),
+            log.body
+        );
     }
 
     Ok(())
+}
+
+fn get_colored_level(level: &Level) -> ColoredString {
+    match level {
+        Level::Trace => level.to_string().bright_black(),
+        Level::Debug => level.to_string().blue(),
+        Level::Info => level.to_string().green(),
+        Level::Warn => level.to_string().yellow(),
+        Level::Error => level.to_string().red(),
+    }
 }
 
 async fn get_deployment_meta(
