+++ common/Cargo.toml
@@ -13,7 +13,6 @@ serde_json = "1.0.79"
 uuid = { version = "0.8.2", features = ["v4", "serde"] }
 rocket = "0.5.0-rc.1"
 anyhow = "1.0.55"
-colored = "2.0"
 lazy_static = "1.4.0"
 log = { version = "0.4", features = ["serde"] }
 chrono = { version = "0.4", features = ["serde"] }
