+++ service/Cargo.toml
@@ -11,6 +11,7 @@ doctest = false
 [dependencies]
 anyhow = "1.0"
 async-trait = "0.1"
+chrono = "0.4"
 libloading = { version = "0.7.3", optional = true }
 log = "0.4"
 sync_wrapper = { version = "0.1", optional = true }
